service-workers/service-worker/mime-sniffing.https.html

<!DOCTYPE html>
<title>Service Worker: MIME sniffing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  const SCOPE = 'resources/blank.html?mime-sniffing';
  const SCRIPT = 'resources/mime-sniffing-worker.js';
  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(SCOPE);

  t.add_cleanup(async() => {
    if (registration)
      await registration.unregister();
    if (frame)
      frame.remove();
  });

  assert_equals(frame.contentWindow.document.body.innerText, 'test');
  const h1 = frame.contentWindow.document.getElementById('testid');
  assert_equals(h1.innerText,'test');
}, 'The response from service worker should be correctly MIME siniffed.');
</script>
